move server.js and index.html files to api folder
