+++ api/index.html
@@ -4,11 +4,11 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
+    <link rel="shortcut icon" type="image/x-icon" href="/favicon.ico"/>
     <title>Anywhere Fitness API</title>
 
     <style>
-      body {
+      body {  
         display: flex;
         justify-content: center;
         align-items: center;
